🫚 Consolodate pparse and parser tests into single module file.
